Add new in-memory training
The idea here is to generalise the framework as it is, and support
a new kind of training whereby you have already loaded all the
data in.

The major work was creating a new InMemoryScene class and new
in-memory training functions that DO NOT take into account
the "Name" of the object. They expect you to *explicitly* provide
the negative training examples.

I also refactored the mixture model slightly and added templated
methods for initialising from a set of positive scenes.